Revert addition of indices to vertex data (won't work)
Previously, I added PositionTextureIndexVertex objects in order to pass
a texture index in to the GPU along with each vertex, planning to use
this to index different petal textures for each rendered instance of a
petal.  However, this won't work since all the instances use the same
vertex data/buffer, and thus the indices can't change on an
instance-by-instance basis (i.e. I'd only be able to change the texture
on a vertex-by-vertex or triangle-by-triangle basis, but that pattern
would be the same across all rendered instances).  I've still left the
code for the PositionTextureIndexVertex in for now, but have reverted
the rendering code to use the PositionTextureVertex structs instead.

graphics.rs:
- Change all PositionTextureIndexVertexes back to
  PositionTextureVertexes for the 5 vertices hard-coded for the textured
  pentagon (in the TEXTURED_PENTAGON_VERTICES constant).
- Similarly revert the vertex buffer layout to be compatible with the
  above change.

graphics/shader.wgsl:
- Revert the textured vertex shader and fragment shader code to use
  vertices that don't include an index.