Add uniform buffer array of petal variants (issue w/ exceeding limit)
NOTE: This code runs, but breaks if I increase the number of petals or
the number of petal variants.  Doing either apparently increases the
number of uniform buffers bound to the fragment shader, and exceeds the
device limit of 12 (thus causing a validation error to crash the
program).  I thought I was passing an array of data in via a single
uniform buffer, but it appears that I'm actually passing an array of
separate uniform buffers in.  I need to figure out how to switch it back
to doing the former so that I can render thousands of petals again like
I could before.

game.rs:
- Add setup code that defines petal variants (the texture index and what
  slice of the texture to use for a single petal).
- Choose random variant indexes for each petal instance.

graphics.rs:
- Rename variables to hold the buffer data and buffer handles for all
  the petal-related data to make them more clear / memorable.
- Add variables for the buffer to hold the petal variants.
- Add debug printing of features / limits of the gpu adapter & device.
- Change the instance setup code to reflect the new indexing of each
  petal to a smaller set of petal variants that defines its texture
  slice.
- Add the new petal variant buffer to the texture binding group.
- Add comments about the uniform buffer binding limit issue I'm running
  into if I increase the number of petals or petal variants.

graphics/gpu_types.rs:
- Add a From<u32> implementation for UniformU32.
- Add a constructor function for the PetalVariant struct.

graphics/shader.wgsl:
- Rename the wrapped_u32 struct to UniformU32 to match the corresponding
  name in my Rust code.
- Rename the PetalTextureInfo struct to PetalVariant to better match the
  naming in the Rust code.
- Add a new uniform buffer array to pass the petal variant indexes in
  through, and change the previous array to be an array of the actual
  variant info (texture index and texture slice data).
- Add comments about the uniform buffer binding limit issue I'm running
  into if I increase the number of petals or petal variants.