Pipe rendered video frames into an ffmpeg process (stack overflows)
Note:  This code for some reason doesn't seem to run from within VSCode
(it crashes when I try to start it using F5), but does seem to run when
run via `cargo run` on the command line.  However, it eventually crashes
after running for a while (after rendering about 4 to 5 minutes worth of
video) due to a stack overflow.  I'm not sure yet what is causing more
and more memory to keep getting allocated on the stack.

graphics.rs:
- Add code to spawn an ffmpeg process in the video thread, and to pipe
  the rendered frame bytes into it.